pyproject.toml
[tool.poetry]
name = "mysaasmp"
version = "0.1.0"
description = "My SaaS MP"
authors = ["WStanley"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

[tool.poetry.group.saas.dependencies]
django = "^5.0"
djangorestframework = "^3.15"
drf-spectacular = "^0.28.0"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9"
celery = "^5.3"
redis = "^5.0"

[tool.poetry.group.bot.dependencies]
aiogram = "^3.22.0"
pydantic-settings = "^2.10.1"
httpx = "^0.27"
pydantic = "^2.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
mypy = "^1.10"
ruff = "^0.6"
datamodel-code-generator = "^0.35.0"

[tool.poetry.group.shared.dependencies]
python-dotenv = "^1.1.1"

[tool.ruff]
line-length = 99
target-version = "py311"
exclude = [
    ".svn", "CVS", ".bzr", ".hg", ".git", "__pycache__", 
    ".tox", ".eggs", "*.egg", ".venv", "env", "venv", "build"
]

[tool.ruff.lint]
select = ["W", "E", "F", "I", "N", "T20", "Q", "T"]
extend-select = ["I", "B", "C", "Q"]
extend-ignore = ["E203", "N818", "E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
docstring-code-line-length = 79
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]

# Основные проверки
warn_return_any = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true

# Для Django и внешних библиотек
ignore_missing_imports = true

# Переопределения для специфичных модулей
[[tool.mypy.overrides]]
module = [
    "django.*",
    "*.migrations.*", 
    "*.admin.*",
    "manage.py",
    "tests.*",
    "*.tests.*",
    "conftest.py",
]
ignore_errors = true

# Или если хотите игнорировать только определенные ошибки в тестах:
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "*.tests.*", 
    "conftest.py",
]
disable_error_code = ["no-untyped-def"]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = ["tests.py", "test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
